🔨Refactor : Refactor badge 2 ~ 6 test2

+++ src/main/java/MTZ/mountainz/domain/member/entity/Member.java
@@ -37,7 +37,7 @@ public class Member {
 	private String region;
 
 	@Column(name = "certificationPoint")
-	private int certificationPoint;
+	private Long certificationPoint;
 
 	@Column(name = "badgeName")
 	private String badgeName;
